helm: Make Hubble mTLS secret/configmap mounts optional
Since #13249, if Hubble is configured to serve the API via mTLS, it is
able to delay starting the server until the certificates are present on
the file system.

As such, we can now mark the mounts for the certificates as optional, as
the we can already start Cilium agent without Hubble server. The Hubble
server will be started automatically once the TLS secrets are ready and
the files are mounted by Kubernetes.

The current Hubble server implementation prints an info log message if
the certificate files are not present after 30 seconds. This allows
users to troubleshoot if the Hubble server is not started due to missing
certificates.

The mounts are still required for Hubble Relay. In contrast to
cilium-agent, it cannot operate without the certificate present.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>